# Simplify `blocks.indexers` and improve performance