--- a/test_xpartition.py
+++ b/test_xpartition.py
@@ -2,167 +2,60 @@ import multiprocessing
 import os
 import string
 
-import pytest
+import dask
 import numpy as np
+import pytest
 import xarray as xr
 
 import xpartition
 
 
 @pytest.mark.parametrize(
-    ("value", "expected"),
-    [(1, True), (np.int32(1), True), (2.0, False), (np.float32(2), False)],
-)
-def test__is_integer(value, expected):
-    result = xpartition._is_integer(value)
-    assert result == expected
-
-
-@pytest.mark.parametrize(
-    ("indexers", "exception"),
+    ("block_indexers", "expected", "exception"),
     [
-        ({"x": 0}, None),
-        ({"x": 4}, None),
-        ({"x": -4}, None),
-        ({"x": slice(1, 2)}, None),
-        ({"x": slice(1, 2, 1)}, None),
-        ({"x": 5}, IndexError),
-        ({"x": -5}, IndexError),
-        ({"x": 2.0}, ValueError),
-        ({"y": 0}, KeyError),
-        ({"x": slice(1, 2, 5)}, NotImplementedError),
+        ({"x": slice(0, 3)}, {"x": slice(0, 6)}, None),
+        ({"x": slice(1, 2)}, {"x": slice(2, 5)}, None),
+        ({"x": slice(-3, -2)}, {"x": slice(0, 2)}, None),
+        ({"x": slice(-3, -1)}, {"x": slice(0, 5)}, None),
+        ({"x": slice(-3, None)}, {"x": slice(0, 6)}, None),
+        ({"x": slice(None, 1)}, {"x": slice(0, 2)}, None),
+        ({"x": slice(0, 10)}, {"x": slice(0, 6)}, None),
+        ({"x": slice(-10, None)}, {"x": slice(0, 6)}, None),
+        ({"x": slice(None, None)}, {"x": slice(0, 6)}, None),
+        ({"x": slice(10, 12)}, {"x": slice(6, 6)}, None),
+        ({"x": slice(2, 1)}, {"x": slice(5, 2)}, None),
+        ({"x": 1}, {"x": slice(2, 5)}, None),
+        ({"x": -1}, {"x": slice(5, 6)}, None),
+        ({"x": -2}, {"x": slice(2, 5)}, None),
+        ({"x": np.int32(2)}, {"x": slice(2, 5)}, None),
+        ({"x": slice(0, 3), "y": 1}, {"x": slice(0, 6), "y": slice(3, 4)}, None),
+        ({"x": 4}, None, IndexError),
+        ({"x": -4}, None, IndexError),
+        ({"z": 1}, None, KeyError),
+        ({"x": slice(None, None, 2)}, None, NotImplementedError),
+        ({"x": 2.0}, None, ValueError),
     ],
     ids=lambda x: f"{x}",
 )
-def test__validate_indexers(indexers, exception):
-    sizes = {"x": 5}
-    if exception is not None:
-        with pytest.raises(exception):
-            xpartition._validate_indexers(indexers, sizes)
+def test_indexers(block_indexers, expected, exception):
+    data = dask.array.zeros((6, 4), chunks=((2, 3, 1), (3, 1)))
+    da = xr.DataArray(data, dims=["x", "y"])
+    if exception is None:
+        result = da.blocks.indexers(**block_indexers)
+        assert result == expected
     else:
-        xpartition._validate_indexers(indexers, sizes)
+        with pytest.raises(exception):
+            da.blocks.indexers(**block_indexers)
 
 
-@pytest.mark.parametrize(
-    ("indexers", "expected"),
-    [
-        ({"x": 5}, {"x": slice(5, 6)}),
-        ({"x": -5}, {"x": slice(-5, -4)}),
-        ({"x": slice(0, 2), "y": 5}, {"x": slice(0, 2), "y": slice(5, 6)}),
-    ],
-    ids=lambda x: f"{x}",
-)
-def test__convert_scalars_to_slices(indexers, expected):
-    result = xpartition._convert_scalars_to_slices(indexers)
-    assert result == expected
+def test_isel():
+    data = dask.array.random.random((6, 4), chunks=((2, 3, 1), (3, 1)))
+    da = xr.DataArray(data, dims=["x", "y"])
 
+    result = da.blocks.isel(x=slice(1, 2), y=1).data.compute()
+    expected = data.blocks[1:2, 1].compute()
 
-@pytest.mark.parametrize(
-    ("block_indexers", "expected"),
-    [
-        ({"x": slice(0, 3)}, {"x": slice(0, 6)}),
-        ({"x": slice(1, 2)}, {"x": slice(2, 5)}),
-        ({"x": slice(-3, -2)}, {"x": slice(0, 2)}),
-        ({"x": slice(-3, -1)}, {"x": slice(0, 5)}),
-        ({"x": slice(-3, None)}, {"x": slice(0, 6)}),
-        ({"x": slice(None, 1)}, {"x": slice(0, 2)}),
-        ({"x": slice(0, 10)}, {"x": slice(0, 6)}),
-        ({"x": slice(-10, None)}, {"x": slice(0, 6)}),
-        ({"x": slice(None, None)}, {"x": slice(0, 6)}),
-        ({"x": slice(10, 12)}, {"x": slice(6, 6)}),
-        ({"x": slice(2, 1)}, {"x": slice(5, 2)}),
-    ],
-    ids=lambda x: f"{x}",
-)
-def test__convert_block_indexers_to_array_indexers(block_indexers, expected):
-    chunks = {"x": (2, 3, 1)}
-    result = xpartition._convert_block_indexers_to_array_indexers(
-        block_indexers, chunks
-    )
-    assert result == expected
-
-
-def _construct_dataarray(shape, chunks, name):
-    dims = list(string.ascii_lowercase[: len(shape)])
-    data = np.random.random(shape)
-    da = xr.DataArray(data, dims=dims, name=name)
-    if chunks is not None:
-        chunks = {dim: chunk for dim, chunk in zip(dims, chunks)}
-        da = da.chunk(chunks)
-    return da
-
-
-SHAPE_AND_CHUNK_PAIRS = [
-    ((5,), (1,)),
-    ((5,), (2,)),
-    ((5,), (5,)),
-    ((2, 5), (1, 1)),
-    ((2, 5), (2, 1)),
-    ((2, 5), (2, 2)),
-    ((2, 5), (2, 4)),
-    ((2, 5), (2, 5)),
-    ((2, 1, 6), (1, 1, 1)),
-    ((2, 1, 6), (1, 1, 2)),
-    ((2, 1, 6), (2, 1, 2)),
-    ((2, 1, 6), (2, 1, 5)),
-    ((2, 3, 4, 5), (1, 1, 1, 1)),
-    ((2, 3, 4, 5), (2, 1, 3, 3)),
-]
-
-
-@pytest.fixture(params=SHAPE_AND_CHUNK_PAIRS, ids=lambda x: str(x))
-def da(request):
-    shape, chunks = request.param
-    name = "foo"
-    return _construct_dataarray(shape, chunks, name)
-
-
-def test_indexers_with_scalars(da):
-    n_blocks = np.product(da.blocks.shape)
-    for i in range(n_blocks):
-        block_indices = np.unravel_index(i, da.blocks.shape)
-        block_indexers = {dim: index for dim, index in zip(da.dims, block_indices)}
-        array_indexers = da.blocks.indexers(**block_indexers)
-        block_data_via_xarray = da.isel(array_indexers).data.compute()
-        block_data_via_dask = da.data.blocks[block_indices].compute()
-        np.testing.assert_array_equal(block_data_via_xarray, block_data_via_dask)
-
-        # Test obtaining the array through blocks.isel
-        block_data_via_xarray = da.blocks.isel(**block_indexers).data.compute()
-        np.testing.assert_array_equal(block_data_via_xarray, block_data_via_dask)
-
-
-@pytest.mark.parametrize(
-    "subset",
-    [
-        {"a": slice(0, 1), "b": slice(0, 1), "c": slice(0, 1), "d": slice(0, 1)},
-        {"a": slice(0, 2), "b": slice(0, 1), "c": slice(0, 1), "d": slice(0, 1)},
-        {"a": slice(0, 1), "b": slice(0, 2), "c": slice(0, 1), "d": slice(0, 1)},
-        {"a": slice(0, 1), "b": slice(0, 1), "c": slice(0, 2), "d": slice(0, 1)},
-        {"a": slice(0, 1), "b": slice(0, 1), "c": slice(0, 1), "d": slice(0, 2)},
-        {"a": slice(1, 2), "b": slice(1, 2), "c": slice(1, 2), "d": slice(1, 2)},
-        {"a": slice(0, 3), "b": slice(0, 1), "c": slice(1, 2), "d": slice(1, 2)},
-        {"a": slice(0, 3), "b": slice(0, 2), "c": slice(0, 1), "d": slice(0, 2)},
-        {"a": slice(0, 4), "b": slice(0, 2), "c": slice(0, 2), "d": slice(0, 2)},
-    ],
-    ids=lambda x: str(x),
-)
-def test_indexers_with_slices(subset):
-    shape = (4, 3, 2, 7)
-    chunks = (1, 2, 1, 5)
-    dims = list(string.ascii_lowercase[: len(shape)])
-    chunks = {dim: chunk for dim, chunk in zip(dims, chunks)}
-    data = np.random.random(shape)
-    da = xr.DataArray(data, dims=dims, name="foo").chunk(chunks)
-
-    indexers = da.blocks.indexers(**subset)
-    xarray_subset = da.isel(indexers).data.compute()
-    dask_subset = da.data.blocks[tuple(s for s in subset.values())].compute()
-    np.testing.assert_array_equal(xarray_subset, dask_subset)
-
-    # Test obtaining the array through blocks.isel
-    xarray_subset = da.blocks.isel(**subset).data.compute()
-    np.testing.assert_array_equal(xarray_subset, dask_subset)
+    np.testing.assert_array_equal(result, expected)
 
 
 @pytest.mark.filterwarnings("ignore:Specified Dask chunks")