This PR greatly simplifies the implementation of `indexers`.  Previously we would compute the array indices of *every single dask block* whenever we would call `indexers`.  This is obviously substantially more work than is needed.  In reality we know which dask blocks we want to compute the indices of in advance, so we can compute those indices alone.  

For arrays without too many dask blocks this was not a big deal, but for arrays with many dask blocks, this was a performance bottleneck.

Resolves #3

## Simple benchmark

This PR results in a greater than 2000x speedup for the following benchmark.  With more chunks I would expect an even greater speedup.

Before:

```
In [1]: import dask.array as darray; import xarray as xr; import xpartition

In [2]: da = xr.DataArray(darray.random.random((10000, 10000), chunks=(100, 100)), dims=["x", "y"])

In [3]: %%timeit
   ...: da.blocks.indexers(x=slice(10, 20), y=slice(25, 50))
   ...: 
   ...: 
21.8 ms ± 195 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After:

```
In [1]: import dask.array as darray; import xarray as xr; import xpartition

In [2]: da = xr.DataArray(darray.random.random((10000, 10000), chunks=(100, 100)), dims=["x", "y"])

In [3]: %%timeit
   ...: da.blocks.indexers(x=slice(10, 20), y=slice(25, 50))
   ...: 
   ...: 
8.67 µs ± 341 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```